requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0
soundfile>=0.12.0
librosa>=0.10.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
gradio>=4.0.0
elevenlabs>=1.0.0
